+++ public/app/features/provisioning/api/utils/getListParams.ts
@@ -1,0 +1,21 @@
+import { parseListOptionsSelector } from '../../../apiserver/client';
+import { ListOptions } from '../../../apiserver/types';
+import { ListRepositoryArg } from '../endpoints.gen';
+
+type ListParams = Omit<ListRepositoryArg, 'fieldSelector' | 'labelSelector'> &
+  Pick<ListOptions, 'labelSelector' | 'fieldSelector'>;
+
+/**
+ * A helper function to remove the watch argument from the queryArg and convert field- and labelSelectors to strings
+ */
+export function getListParams(queryArg: ListParams) {
+  if (!queryArg) {
+    return {};
+  }
+  const { fieldSelector, labelSelector, watch, ...params } = queryArg;
+  return {
+    fieldSelector: fieldSelector ? parseListOptionsSelector(fieldSelector) : undefined,
+    labelSelector: labelSelector ? parseListOptionsSelector(labelSelector) : undefined,
+    ...params,
+  };
+}
